% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_CDTsamples_convergence.R
\name{check_CDTsamples_convergence}
\alias{check_CDTsamples_convergence}
\title{check_CDTsamples_convergence TITLE}
\usage{
check_CDTsamples_convergence(CDTsamples)
}
\arguments{
\item{CDTsamples}{XXXXXXX.}
}
\value{
TRUE if the Gelman Rubin test for convergence was successful, FALSE otherwise
}
\description{
\code{check_CDTsamples_convergence} DESCRIPTION TO COME.
}
\details{
{
XXXXXXX. 
}
}
\examples{
## XXXXXXX.
}
\references{
XXXXXXX.
}
\seealso{
XXXXXXX.
}
\author{
XXXXXXX.
}
