% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mv_epiestim_inputs.R
\name{check_incidence}
\alias{check_incidence}
\title{Check incidence input for MV-EpiEstim}
\usage{
check_incidence(incid)
}
\arguments{
\item{incid}{a multidimensional array containing values of the incidence
for each time step (1st dimension), location (2nd dimension) and
pathogen/strain/variant (3rd dimension)}
}
\value{
Silently returns TRUE if the checks are passed, otherwise throws an
error.
}
\description{
Check that incid is a 3-dimensional array with non-negatuve enteries.
}
\author{
Sangeeta Bhatia
}
\keyword{internal}
