% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mv_epiestim_inputs.R
\name{check_mcmc_control}
\alias{check_mcmc_control}
\title{Check MCMC control parameters for MV-EpiEstim}
\usage{
check_mcmc_control(mcmc_control)
}
\arguments{
\item{mcmc_control}{a list of default MCMC control parameters, obtained by
default from \code{\link[=default_mcmc_controls]{default_mcmc_controls()}}}
}
\value{
Silently returns TRUE if the checks are passed, otherwise throws an
error.
}
\description{
Check that (1) mcmc_control is a list of the correct format, (2) n_iter, burnin
and thin are positive integers and (3) n_iter is greater than burnin + thin.
}
\author{
Sangeeta Bhatia
}
\keyword{internal}
