% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mv_epiestim_inputs.R
\name{check_priors}
\alias{check_priors}
\title{Check priors  for MV-EpiEstim}
\usage{
check_priors(priors)
}
\arguments{
\item{priors}{a list of prior parameters (shape and scale of a gamma
distribution) for epsilon and R; can be obtained from the function
\code{\link[=default_priors]{default_priors()}}. The prior for R is assumed to be the same for all
time steps and all locations}
}
\value{
Silently returns TRUE if the checks are passed, otherwise throws an
error.
}
\description{
Check that priors is a list of the correct format and that the mean of the
prior for epsilon is 1 (as currently only priors with mean of epsilon equal
to 1 is supported).
}
\author{
Sangeeta Bhatia
}
