% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mv_epiestim_inputs.R
\name{check_seed}
\alias{check_seed}
\title{Validate seed input for MV-EpiEstim}
\usage{
check_seed(seed)
}
\arguments{
\item{seed}{a numeric value used to fix the random seed}
}
\value{
Silently returns TRUE if the checks are passed, otherwise throws an error.
}
\description{
Validate that the seed is not null and is numeric
}
\author{
Sangeeta Bhatia
}
