% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_mv_epiestim_inputs.R
\name{check_t_min_t_max}
\alias{check_t_min_t_max}
\title{Validate t_min and t_max inputs for MV-EpiEstim}
\usage{
check_t_min_t_max(t_min, t_max, incid)
}
\arguments{
\item{t_min}{an integer > 1 giving the minimum time step to consider in the
estimation.
If \code{NULL}, \code{t_min} is calculated using the function \code{\link[=compute_si_cutoff]{compute_si_cutoff()}}
which gets the maximum (across all variants) of the 95th percentile of the
SI distribution.}

\item{t_max}{an integer > \code{t_min} and <= \code{nrow(incid)} giving the maximum time
step to consider in the estimation. Default value is \code{nrow(incid)}.}

\item{incid}{a multidimensional array containing values of the incidence
for each time step (1st dimension), location (2nd dimension) and
pathogen/strain/variant (3rd dimension)}
}
\value{
Silently returns TRUE if the checks are passed, otherwise throws an
error.
}
\description{
Check that (1) t_min and t_max are integers, (2) that they are >= 2, that
they are <= nrow(incid) and (3) t_min is not greater than t_max.
}
\author{
Sangeeta Bhatia
}
\keyword{internal}
