% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{compute_lambda}
\alias{compute_lambda}
\title{Compute the overall infectivity}
\usage{
compute_lambda(incid, si_distr)
}
\arguments{
\item{incid}{a list (as obtained from function `process_I_multivariant`)
of two multidimensional arrays ("local" and "imported")
containing values of the incidence
for each time step (1st dimension), location (2nd dimension) and
pathogen/strain/variant (3rd dimension)}

\item{si_distr}{a matrix where each column contains the probability mass
function for the discrete serial interval for each of the
pathogen/strain/variants, starting with the probability mass function
for day 0 in the first row, which should be 0. Each column in the matrix
should sum to 1}
}
\value{
a multidimensional array containing values of the overall
  infectivity for each time step (1st dimension), location (2nd dimension)
  and pathogen/strain/variant (3rd dimension). The overall infectivity for
  a given location and pathogen/strain/variant represents the sum of
  the incidence for that location and that pathogen/strain/variant at all
  previous time steps, weighted by the current infectivity of those
  past incident cases. Pre-calculating the overall infectivity makes the
  algorithm much faster
}
\description{
Compute the overall infectivity
}
\examples{

n_v <- 2
n_loc <- 3 # 3 locations
T <- 100 # 100 time steps
priors <- default_priors()
# constant incidence 10 per day everywhere
incid <- array(10, dim = c(T, n_loc, n_v))
incid <- process_I_multivariant(incid)
# arbitrary serial interval, same for both variants
w_v <- c(0, 0.2, 0.5, 0.3)
si_distr <- cbind(w_v, w_v)
lambda <- compute_lambda(incid, si_distr)
}
