% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{compute_si_cutoff}
\alias{compute_si_cutoff}
\title{Index before which at most a given probability
mass is captured}
\usage{
compute_si_cutoff(si_distr, miss_at_most = 0.05)
}
\arguments{
\item{si_distr}{a matrix of discretised probability distributions
(each column sums to 1). This could be privided by the user if using
non_parametric serial interval distributions, or could be generated internally
by the function \code{\link[=estimate_advantage]{estimate_advantage()}} if using parametric serial interval
distribution specification.}

\item{miss_at_most}{numeric. Probability mass in the tail of the SI distribution}
}
\value{
integer
}
\description{
Across a matrix of discretised probability distributions
(see \code{\link[=estimate_advantage]{estimate_advantage()}}) this function returns the largest index
(across all columns) such that the cumulative probability mass before index is
\code{1 - miss_at_most}.
}
\author{
Sangeeta Bhatia
}
