% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{compute_t_min}
\alias{compute_t_min}
\title{Compute the smallest index at which joint estimation should start}
\usage{
compute_t_min(incid, si_distr, miss_at_most)
}
\arguments{
\item{incid}{a multidimensional array containing values of the incidence
for each time step (1st dimension), location (2nd dimension) and
pathogen/strain/variant (3rd dimension)}

\item{si_distr}{a matrix of discretised probability distributions
(each column sums to 1). This could be privided by the user if using
non_parametric serial interval distributions, or could be generated internally
by the function \code{\link[=estimate_advantage]{estimate_advantage()}} if using parametric serial interval
distribution specification.}

\item{miss_at_most}{numeric. Probability mass in the tail of the SI distribution}
}
\value{
integer
}
\description{
Unless specified by the user, \code{t_min} in \code{\link[=estimate_advantage]{estimate_advantage()}} is computed
as the sum of two indices:
}
\details{
\itemize{
\item the first day of non-zero incidence across all locations, computed using
\code{\link[=first_nonzero_incid]{first_nonzero_incid()}}
\item the 95th percentile of the probability mass function of the SI distribution
across all variants computed using \code{\link[=compute_si_cutoff]{compute_si_cutoff()}}
}
}
\author{
Sangeeta Bhatia
}
