% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_distr_name_for_mcmc}
\alias{convert_distr_name_for_mcmc}
\title{Convert EpiEstim distribution names to those used by coarsedatatools}
\usage{
convert_distr_name_for_mcmc(distr)
}
\arguments{
\item{distr}{A string with the name of the distribution as provided by the user.}
}
\value{
A string with the converted distribution name.
}
\description{
coarsedatatools uses abberviated names for distributions e.g. "G" for gamma etc
To provide a smooth user exper experience, we convert the more descriptive names.
This function performs the user-provided names to the abberviated ones.
}
\author{
Sangeeta Bhatia
}
\keyword{internal}
