% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{default_mcmc_controls}
\alias{default_mcmc_controls}
\title{Set default for MCMC control}
\usage{
default_mcmc_controls()
}
\value{
a list of default MCMC control parameters, containing:

- n_iter: the number if iterations of the MCMC to perform

- burnin: the burnin to use; MCMC iterations will only be recorded after
  the burnin

- thin: MCMC iterations will only be recorded after
  the burnin and every `thin` iteration

  Values can then be manually be edited as in the examples below.
}
\description{
Set default for MCMC control
}
\examples{
mcmc_control<- default_mcmc_controls()
# change to run for 10 times longer
mcmc_control$n_iter <- mcmc_control$n_iter * 10

}
