% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{default_priors}
\alias{default_priors}
\title{Set default for Gamma priors}
\usage{
default_priors()
}
\value{
a list of default parameters for the priors.
  Values can then be manually be edited as in the examples below.
  Users could use functions `epitrix::gamma_shapescale2mucv` and
  `epitrix::gamma_mucv2shapescale` to set the shape and scale corresponding
  to the desired prior mean and coefficient of variation.
}
\description{
Set default for Gamma priors
}
\examples{
priors <- default_priors()
# change the prior for R to have a mean of 3
priors$R$shape <- 3

}
