% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{DiscrSI}
\alias{DiscrSI}
\title{Function to ensure compatibility with EpiEstim versions <2.0}
\usage{
DiscrSI(k, mu, sigma)
}
\arguments{
\item{k}{Positive integer, or vector of positive integers for which the
discrete distribution is desired.}

\item{mu}{A positive real giving the mean of the Gamma distribution.
Must be greater than \code{shift} (1 by default).}

\item{sigma}{A non-negative real giving the standard deviation of the Gamma
distribution.}
}
\description{
Please only use for compatibility;
Prefer the new \code{\link[=discr_si]{discr_si()}} function instead
}
