% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiEstim-package.R
\docType{package}
\name{EpiEstim-package}
\alias{EpiEstim}
\alias{EpiEstim-package}
\title{EpiEstim: Estimate Time Varying Reproduction Numbers from Epidemic Curves}
\description{
The EpiEstim package provides tools to Tools to quantify transmissibility throughout
an epidemic from the analysis of time series of incidence.
}
\section{Bibliography}{

A BibTeX file of  papers that describe the methodology underlying EpiEstim is available via
\code{system.file("epiestimpapers.bib", package = "EpiEstim")}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrc-ide/EpiEstim}
  \item Report bugs at \url{https://github.com/mrc-ide/EpiEstim/issues}
}

}
\author{
\strong{Maintainer}: Anne Cori \email{a.cori@imperial.ac.uk} (\href{https://orcid.org/0000-0002-8443-9162}{ORCID})

Other contributors:
\itemize{
  \item Simon Cauchemez [contributor]
  \item Neil M. Ferguson (\href{https://orcid.org/0000-0002-1154-8093}{ORCID}) [contributor]
  \item Christophe Fraser (\href{https://orcid.org/0000-0003-2399-9657}{ORCID}) [contributor]
  \item Elisabeth Dahlqwist (\href{https://orcid.org/0000-0001-5797-6803}{ORCID}) [contributor]
  \item P. Alex Demarsh [contributor]
  \item Thibaut Jombart (\href{https://orcid.org/0000-0003-2226-8692}{ORCID}) [contributor]
  \item Zhian N. Kamvar (\href{https://orcid.org/0000-0003-1458-7108}{ORCID}) [contributor]
  \item Justin Lessler (\href{https://orcid.org/0000-0002-9741-8109}{ORCID}) [contributor]
  \item Shikun Li [contributor]
  \item Jonathan A. Polonsky (\href{https://orcid.org/0000-0002-8634-4255}{ORCID}) [contributor]
  \item Jake Stockwin [contributor]
  \item Robin Thompson (\href{https://orcid.org/0000-0001-8545-5212}{ORCID}) [contributor]
  \item Rolina van Gaalen \email{rolina.van.gaalen@rivm.nl} [contributor]
  \item Rebecca Nash \email{r.nash@imperial.ac.uk} (\href{https://orcid.org/0000-0002-5213-4364}{ORCID}) [contributor]
  \item Sangeeta Bhatia (\href{https://orcid.org/0000-0001-6525-101X}{ORCID}) [contributor]
  \item Jack Wardle (\href{https://orcid.org/0000-0002-5363-1653}{ORCID}) [contributor]
  \item Andrea Brizzi (\href{https://orcid.org/0000-0002-5639-8260}{ORCID}) [contributor]
}

}
\keyword{internal}
