% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_agg.R
\name{estimate_R_agg}
\alias{estimate_R_agg}
\title{Estimate instantaneous reproduction number from coarsely aggregated data}
\usage{
estimate_R_agg(
  incid,
  dt = 7L,
  dt_out = 7L,
  iter = 10L,
  tol = 0.000001,
  recon_opt = "naive",
  config = make_config(),
  method = c("non_parametric_si", "parametric_si"),
  grid = list(precision = 0.001, min = -1, max = 1),
  agg_dates = NULL,
  date_convention = NULL
)
}
\arguments{
\item{incid}{Aggregated incidence data One of the following
\itemize{
\item A vector (or a dataframe with a single column) of non-negative integers
containing the incidence time series; these can be aggregated at any time
unit as specified by argument \code{dt}
\item A dataframe of non-negative integers with either i) \code{incid$I}
containing the total incidence, or ii) two columns, so that
\code{incid$local} contains the incidence of cases due to local transmission
and \code{incid$imported} contains the incidence of imported cases (with
\code{incid$local + incid$imported} the total incidence). If the dataframe
contains a column \code{incid$dates}, this is used for plotting.
\code{incid$dates} must contains only dates in a row.
\item An object of class \code{\link[incidence:incidence]{incidence::incidence()}}
}

Note that the cases from the first time step are always all assumed to be
imported cases.
\itemize{
\item An object of class incidence2::incidence()
}}

\item{dt}{length of temporal aggregations of the incidence data. This should
be an integer or vector of integers. If a vector, this will be recycled. For
example, \code{dt = c(3L, 4L)} would correspond to alternating incidence
aggregation windows of 3 and 4 days. The default value is 7 time units
(typically days) - see details.}

\item{dt_out}{length of the sliding windows used for R estimates (numeric,
7 time units (typically days)  by default). Only used if \code{dt > 1}; in
this case this will supersede config$t_start and config$t_end, see \code{\link[=estimate_R]{estimate_R()}}.}

\item{iter}{number of iterations of the EM algorithm (integer, 10 by default)}

\item{tol}{tolerance used in the convergence check (numeric, 1e-6 by default)}

\item{recon_opt}{one of "naive" or "match" (see details).}

\item{config}{an object of class \code{estimate_R_config}, as returned by
function \code{\link[=make_config]{make_config()}}.}

\item{method}{one of "non_parametric_si" or "parametric_si" (see details).}

\item{grid}{named list containing \code{precision}, \code{min}, and \code{max} which are used to
define a grid of growth rate parameters that are used inside the EM algorithm
(see details). We recommend using the default values.}

\item{agg_dates}{dates corresponding to each aggregated incidence count,
supplied as a vector of class \code{Date}. If supplied, \code{date_convention}
must also be specified. If \code{NULL} (default), no date information is
added to the output.}

\item{date_convention}{One of \code{"start"} or \code{"end"}, specifying whether the
dates supplied in \code{incid$dates} correspond to the first or last day of each
aggregation window. Must be specified if incidence are temporally
aggregated (\code{dt > 1}) and \code{incid$dates} is supplied.
For example, for weekly epi-week data where the date label
corresponds to the first day of the week (e.g. Monday), use \code{"start"}. For
data reported multiple times per week where the date corresponds to the
reporting date (i.e. the last day of the accumulation window), use \code{"end"}.
Ignored if no dates are supplied.}
}
\value{
An object of class \code{\link[=estimate_R]{estimate_R()}}, with components:
\itemize{
\item \code{R}: a dataframe containing the times of start and end of each time window
considered; the posterior mean, std, and 0.025, 0.05, 0.25, 0.5, 0.75, 0.95,
0.975 quantiles of the reproduction number for each time window.
\item \code{method}: the method used to estimate R, one of "non_parametric_si",
"parametric_si", "uncertain_si", "si_from_data" or "si_from_sample"
\item \code{si_distr}: a vector or dataframe (depending on the method) containing
the discrete serial interval distribution(s) used for estimation
\item \code{SI.Moments}: a vector or dataframe (depending on the method)
containing the mean and std of the discrete serial interval distribution(s)
used for estimation
\item \code{I}: the time series of daily incidence reconstructed by the EM algorithm.
For the initial incidence that cannot be reconstructed (e.g. the first aggregation
window and aggregation windows where incidence is too low to estimate Rt - see
details) then the incidence returned will be the naive disaggregation of the
incidence data used to initialise the EM algorithm.
\item \code{I_local}: the time series of incidence of local cases (so that
\code{I_local + I_imported = I})
\item \code{I_imported}: the time series of incidence of imported cases (so that
\code{I_local + I_imported = I})
\item \code{dates}: a vector of dates corresponding to the incidence time series
}
}
\description{
Estimate instantaneous reproduction number from coarsely aggregated data
}
\details{
Estimation of the time-varying reproduction number from temporally-aggregated
incidence data. For full details about how Rt is estimated within this
bayesian framework, see details in \code{\link[=estimate_R]{estimate_R()}}.

Here, an expectation maximisation (EM) algorithm is used to reconstruct daily
incidence from data that is provided on another timescale. In addition to the
usual parameters required in \code{\link[=estimate_R]{estimate_R()}}, \code{estimate_R_agg()}
requires that the user specify two additional parameters: \code{dt} and
\code{dt_out}. There are two other parameters that the user may modify, \code{iter} and
\code{grid}, however we recommend that the default values are used.
\itemize{
\item \code{dt} is the length of the temporal aggregations of the incidence data in
days. This can be a single integer if the aggregations do not change. E.g.
for weekly data, the user would supply \code{dt = 7L}. If the aggregation windows
vary in length, a vector of integers can be supplied. If the aggregations
have a repeating pattern, for instance, if incidence is always reported three
times per week on the same day of the week, you can supply a vector such as:
\code{dt = c(2L,2L,3L)}. If the aggregations change over time, or do not have a
repeating pattern, the user can supply a full vector of aggregations matching
the length of the incidence data supplied.
\item \code{dt_out} is the length of the sliding window used to estimate Rt from the
reconstructed daily data. By default, Rt estimation uses weekly sliding time
windows, however, we recommend that the sliding window is at least equal to
the length of the longest aggregation window (\code{dt}) in the data.
\item \code{recon_opt} specifies how to handle the initial incidence data that cannot
be reconstructed by the EM algorithm (e.g. the incidence data for the aggregation
window that precedes the first aggregation window that R can be estimated for).
If \code{"naive"} is chosen, the naive disaggregation of the incidence data will be
kept. If \code{"match"} is chosen, the incidence in the preceding aggregation window
will be reconstructed by assuming that the growth rate matches that of the first
estimation window. This is \code{"naive"} by default.
\item \code{iter} is the number of iterations of the EM algorithm used to reconstruct
the daily incidence data. By default, \code{iter = 10L}, which has been demonstrated
to exceed the number of iterations necessary to reach convergence in simulation
studies and analysis of real-world data by the package authors (manuscript in
prep).
\item \code{grid} is a named list containing \code{precision}, \code{min}, and \code{max} which are
used to define a grid of growth rate parameters used inside the EM algorithm.
The grid is used to convert reproduction number estimates for each aggregation
of incidence data into growth rates, which are then used to reconstruct the
daily incidence data assuming exponential growth. The grid will auto-adjust
if it is not large enough, so we recommend using the default values.
}

There are three stages of the EM algorithm:
\itemize{
\item Initialisation. The EM algorithm is initialised with a naive disaggregation
of the incidence data. For example, if there were 70 cases over the course of a
week, this would be naively split into 10 cases per day.
\item Expectation. The reproduction number is estimated for each aggregation
window, except for the first aggregation window (as there is no past incidence
data). This means that the earliest the incidence reconstruction can start is
at least the first day of the second aggregation window. Additionally, if the
disaggregated incidence in subsequent aggregation windows is too low to estimate
the reproduction number, this will mean that the reconstruction will not start
until case numbers are sufficiently high.
\item Maximisation. The reproduction number estimates are then translated
into growth rates for each aggregation window (Wallinga & Lipsitch, 2007) and
used to reconstruct daily incidence data assuming exponential growth. The daily
incidence is adjusted by a constant to ensure that if the daily incidence were
to be re-aggregated, it would still sum to the original aggregated totals. The
expectation and maximisation steps repeat iteratively until convergence.
}

The daily incidence that is reconstructed after the final iteration of the EM
algorithm is then used to estimate Rt using the same process as the original
\code{\link[=estimate_R]{estimate_R()}}` function, with sliding weekly time windows used as the default.
}
\examples{
## Example for constant aggregation windows e.g. weekly reporting

# Load data on SARS in 2003
data("SARS2003")

# this is daily data, but for this example we will aggregate it to weekly 
# counts using the `aggregate_inc()` function
incid <- SARS2003$incidence
dt <- 7L
weekly_incid <- aggregate_inc(incid, dt)
si_distr <- SARS2003$si_distr

# estimate Rt using the default parameters (method "non_parametric_si")
method <- "non_parametric_si"
config <- make_config(list(si_distr = si_distr))
res_weekly <- estimate_R_agg(incid = weekly_incid, 
                            dt = 7L, # aggregation window of the data
                            dt_out = 7L, # desired sliding window length
                            iter = 10L,
                            config = config,
                            method = method,
                            grid = list(precision = 0.001, min = -1, max = 1))

# Plot the result
plot(res_weekly)


## Example using repeating vector of aggregation windows e.g. for consistent 
## reporting 3 times a week

# Using the SARS data again
data("SARS2003")

# For this example we will pretend data is being reported three times a week,
# in 2-day, 2-day, 3-day aggregations
incid <- SARS2003$incidence
dt <- c(2L,2L,3L)
agg_incid <- aggregate_inc(incid, dt)
si_distr <- SARS2003$si_distr 

# estimate Rt using the default parameters (method "non_parametric_si")
method <- "non_parametric_si"
config <- make_config(list(si_distr = si_distr))
res_agg <- estimate_R_agg(incid = agg_incid, 
                            dt = c(2L,2L,3L), # aggregation windows of the data
                            dt_out = 7L, # desired sliding window length
                            iter = 10L,
                            config = config,
                            method = method,
                            grid = list(precision = 0.001, min = -1, max = 1))

# Plot the result
plot(res_agg)

## Example using full vector of aggregation windows

dt <- rep(c(2L,2L,3L), length.out=length(agg_incid))
si_distr <- SARS2003$si_distr 

# estimate Rt using the default parameters (method "non_parametric_si")
method <- "non_parametric_si"
config <- make_config(list(si_distr = si_distr))
res_agg <- estimate_R_agg(incid = agg_incid, 
                            dt = dt, # aggregation windows of the data
                            dt_out = 7L, # desired sliding window length
                            iter = 10L,
                            config = config,
                            method = method,
                            grid = list(precision = 0.001, min = -1, max = 1))

# Plot the result
plot(res_agg)
}
\references{
Nash RK, Cori A, Nouvellet P. Estimating the epidemic reproduction number
from temporally aggregated incidence data: a statistical modelling approach
and software tool. \doi{https://doi.org/10.1371/journal.pcbi.1011439}.

Wallinga & Lipsitch. How generation intervals shape the relationship between
growth rates and reproductive numbers (Proc Biol Sci 2007).
}
\seealso{
\code{\link[=estimate_R]{estimate_R()}} for details of the core function
}
\author{
Rebecca Nash and Anne Cori
}
