% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{EstimateR}
\alias{EstimateR}
\title{Function to ensure compatibility with EpiEstim versions <2.0}
\usage{
EstimateR(
  I,
  T.Start,
  T.End,
  method = c("NonParametricSI", "ParametricSI", "UncertainSI"),
  n1 = NULL,
  n2 = NULL,
  Mean.SI = NULL,
  Std.SI = NULL,
  Std.Mean.SI = NULL,
  Min.Mean.SI = NULL,
  Max.Mean.SI = NULL,
  Std.Std.SI = NULL,
  Min.Std.SI = NULL,
  Max.Std.SI = NULL,
  SI.Distr = NULL,
  Mean.Prior = 5,
  Std.Prior = 5,
  CV.Posterior = 0.3,
  plot = FALSE,
  leg.pos = "topright"
)
}
\arguments{
\item{I}{see `incid` in [estimate_R()]}

\item{T.Start}{see `config$t_start` from [make_config()] used in [estimate_R()]}

\item{T.End}{see `config$t_end` from [make_config()] used in [estimate_R()]}

\item{method}{see `method` in [estimate_R()] (but [EstimateR()] uses CamelCase 
where [estimate_R()] uses snake_case for the method names)}

\item{n1}{see `n1` in [estimate_R()]}

\item{n2}{see `n2` in [estimate_R()]}

\item{Mean.SI}{see `config$mean_si` from [make_config()] used in [estimate_R()]}

\item{Std.SI}{see `config$std_si` from [make_config()] used in [estimate_R()]}

\item{Std.Mean.SI}{see `config$std_mean_si` from [make_config()] used in 
[estimate_R()]}

\item{Min.Mean.SI}{see `config$min_mean_si` from [make_config()] used in 
[estimate_R()]}

\item{Max.Mean.SI}{see `config$max_mean_si` from [make_config()] used in 
[estimate_R()]}

\item{Std.Std.SI}{see `config$std_std_si` from [make_config()] used in 
[estimate_R()]}

\item{Min.Std.SI}{see `config$min_std_si` from [make_config()] used in 
[estimate_R()]}

\item{Max.Std.SI}{see `config$max_std_si` from [make_config()] used in 
[estimate_R()]}

\item{SI.Distr}{see `config$si_distr` from [make_config()] used in 
[estimate_R()]}

\item{Mean.Prior}{see `config$mean_prior` from [make_config()] used in 
[estimate_R()]}

\item{Std.Prior}{see `config$std_prior` from [make_config()] used in 
[estimate_R()]}

\item{CV.Posterior}{see `config$cv_posterior` from [make_config()] used in 
[estimate_R()]}

\item{plot}{Not used anymore, only there for compatibility}

\item{leg.pos}{Not used anymore, only there for compatibility}
}
\description{
Please only use for compatibility;
Prefer the new [estimate_R()] function instead.
}
