% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{first_nonzero_incid}
\alias{first_nonzero_incid}
\title{Get the first day of non-zero incidence across
all variants and locations.}
\usage{
first_nonzero_incid(incid)
}
\arguments{
\item{incid}{a multidimensional array containing values of the incidence
for each time step (1st dimension), location (2nd dimension) and
pathogen/strain/variant (3rd dimension)}
}
\value{
integer
}
\description{
Get the first day of non-zero incidence across
all variants and locations.
}
\details{
For each variant, find the first day of
non-zero incidence. The maximum of these
is the smallest possible point at which
estimation can begin.
}
\author{
Sangeeta Bhatia
}
