% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvepiestim_precompute.R
\name{get_scale_epsilon}
\alias{get_scale_epsilon}
\title{Compute scale of epsilon posterior}
\usage{
get_scale_epsilon(R, lambda, priors, t)
}
\arguments{
\item{R}{a matrix with dimensions containing values of the instantaneous
reproduction number for each time step (row) and location (column), for
the reference pathogen/strain/variant}

\item{lambda}{a multidimensional array containing values of the overall
infectivity for each time step (1st dimension), location (2nd dimension)
and pathogen/strain/variant (3rd dimension). The overall infectivity for
a given location and pathogen/strain/variant represents the sum of
the incidence for that location and that pathogen/strain/variant at all
previous time steps, weighted by the current infectivity of those
past incident cases. It can be calculated from the incidence \code{incid} and
the distribution of the serial interval using function \code{\link[=compute_lambda]{compute_lambda()}}}

\item{priors}{a list of prior parameters (shape and scale of a gamma
distribution) for epsilon and R; can be obtained from the function
\code{\link[=default_priors]{default_priors()}}. The prior for R is assumed to be the same for all
time steps and all locations}

\item{t}{integer vector of time indices used for estimation.}
}
\value{
a numeric vector of scales for each non-reference variant.
}
\description{
Compute scale of epsilon posterior
}
