% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{get_shape_R_flat}
\alias{get_shape_R_flat}
\title{Precompute shape of posterior distribution for R}
\usage{
get_shape_R_flat(incid, priors, t_min = 2L, t_max = nrow(incid))
}
\arguments{
\item{incid}{a multidimensional array containing values of the (local)
incidence
for each time step (1st dimension), location (2nd dimension) and
pathogen/strain/variant (3rd dimension)}

\item{priors}{a list of prior parameters (shape and scale of a gamma
distribution) for epsilon and R; can be obtained from the function
`default_priors`. The prior for R is assumed to be the same for all
time steps and all locations}

\item{t_min}{an integer >1 giving the minimum time step to consider in the
estimation. Default value is 2 (as the estimation is conditional on
observations at time step 1 and can therefore only start at time step 2).}

\item{t_max}{an integer >`t_min` and <=`nrow(incid)` giving the maximum time
step to consider in the estimation. Default value is `nrow(incid)`.}
}
\value{
a vector of the shape of the posterior distribution of R for 
  each time step t and each location l 
  (stored in element (l-1)*(t_max - t_min + 1) + t of the vector)
}
\description{
Precompute shape of posterior distribution for R
}
\examples{

n_v <- 2
n_loc <- 3 # 3 locations
T <- 100 # 100 time steps
priors <- default_priors()
# constant incidence 10 per day everywhere
incid <- array(10, dim = c(T, n_loc, n_v))
get_shape_R_flat(incid, priors)

}
