% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mcmc_control.R
\name{make_mcmc_control}
\alias{make_mcmc_control}
\title{Create list of MCMC control parameters}
\usage{
make_mcmc_control(
  burnin = 3000,
  thin = 10,
  seed = as.integer(Sys.time()),
  init_pars = NULL
)
}
\arguments{
\item{burnin}{Not used. Kept so that existing code runs.}

\item{thin}{Not used. Kept so that existing code runs.}

\item{seed}{An integer used as the seed for the random number generator
when drawing the sample of serial interval distributions; useful to get
reproducible results.}

\item{init_pars}{vector of size 2 corresponding to the initial values of
parameters to use for the SI distribution. This is the shape and scale for
all but the lognormal distribution, for which it is the meanlog and
sdlog.}
}
\value{
An object of class \code{estimate_R_mcmc_control} with components
\code{burnin}, \code{thin}, \code{seed}, \code{init_pars}. This can be
used as an argument of function \code{\link[=make_config]{make_config()}}.
}
\description{
Creates a list of control parameters to be used in
\code{config$mcmc_control}, where \code{config} is an argument of the
\code{\link[=estimate_R]{estimate_R()}} function. This is used to configure the estimation of the
serial interval within \code{\link[=estimate_R]{estimate_R()}} (with method "si_from_data").
The serial interval is estimated by maximum likelihood, so \code{burnin} and
\code{thin} are not used, and a warning is given if they differ from their
defaults.
}
\details{
The argument \code{si_data}, should be a dataframe with 5 columns:
\itemize{
\item \code{EL}: the lower bound of the symptom onset date of the infector (given as
an integer)
\item \code{ER}: the upper bound of the symptom onset date of the infector (given as
an integer). Should be such that \code{ER >= EL}
\item \code{SL}: the lower bound of the symptom onset date of the infected
individual (given as an integer)
\item \code{SR}: the upper bound of the symptom onset date of the infected
individual (given as an integer). Should be such that \code{SR >= SL}
\item \code{type} (optional): can have entries 0, 1, or 2, corresponding to doubly
interval-censored, single interval-censored or exact observations,
respectively, see Reich et al. Statist. Med. 2009. If not specified, this
will be automatically computed from the dates
}

Assuming a given parametric distribution for the serial interval distribution
(specified in \code{si_parametric_distr}),
the serial interval is estimated directly from these data by maximum
likelihood using \code{\link[primarycensored:fitdistdoublecens]{primarycensored::fitdistdoublecens()}}.
}
\examples{
\dontrun{
## Note the following examples estimate the serial interval
## distribution from data, so they may take a few seconds to run

## load data on rotavirus
data("MockRotavirus")

## estimate the reproduction number (method "si_from_data")
mcmc_seed <- 1
mcmc_control <- make_mcmc_control(seed = mcmc_seed)

incid <- MockRotavirus$incidence
method <- "si_from_data"
overall_seed <- 2
config <- make_config(incid = incid, 
                     si_parametric_distr = "G",
                     mcmc_control = mcmc_control,
                     n1 = 500
                     n2 = 50,
                     seed = overall_seed)

R_si_from_data <- estimate_R(incid,
                            method = method,
                            si_data = MockRotavirus$si_data,
                            config = config)
}
}
