% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{si_from_data_valid_distrs}
\alias{si_from_data_valid_distrs}
\title{Distribution names valid when estimating the serial interval from data}
\usage{
si_from_data_valid_distrs(dist)
}
\arguments{
\item{dist}{A string with the name of a distribution.}
}
\value{
A two element list - the first element is a flag \code{is_dist_valid}
indicating whether the passed distribution is valid. Te second element
\code{all_valid_distrs} is a character vector with the valid distribution names.
}
\description{
When using the si_from_data method, the serial interval is estimated with
primarycensored for one of a set of named distributions. This function
returns the valid distribution names as used by EpiEstim.
}
\author{
Sangeeta Bhatia
}
