% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{WT}
\alias{WT}
\title{Function to ensure compatibility with EpiEstim versions <2.0}
\usage{
WT(
  I,
  T.Start,
  T.End,
  method = c("NonParametricSI", "ParametricSI"),
  Mean.SI = NULL,
  Std.SI = NULL,
  SI.Distr = NULL,
  nSim = 10,
  plot = FALSE,
  leg.pos = "topright"
)
}
\arguments{
\item{I}{see `incid` in [wallinga_teunis()]}

\item{T.Start}{see `config$t_start` in [wallinga_teunis()]}

\item{T.End}{see `config$t_end` in [wallinga_teunis()]}

\item{method}{see `method` in [wallinga_teunis()] (but WT uses CamelCase 
where wallinga_teunis uses snake_case for the method names)}

\item{Mean.SI}{see `config$mean_si` in [wallinga_teunis()]}

\item{Std.SI}{see `config$std_si` in [wallinga_teunis()]}

\item{SI.Distr}{see `config$si_distr` in [wallinga_teunis()]}

\item{nSim}{see `config$n_sim` in [wallinga_teunis()]}

\item{plot}{Not used anymore, only there for compatibility}

\item{leg.pos}{Not used anymore, only there for compatibility}
}
\description{
Please only use for compatibility;
Prefer the new [wallinga_teunis()] function instead
}
