% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_inc.R
\name{aggregate_inc}
\alias{aggregate_inc}
\title{Aggregating daily incidence to longer time windows}
\usage{
aggregate_inc(incid, dt = 7L)
}
\arguments{
\item{incid}{a vector of daily incidence values}

\item{dt}{a positive integer indicating the length of the desired aggregation window}
}
\value{
a vector of incidence values, aggregated to dt
}
\description{
Aggregating daily incidence to longer time windows
}
\examples{
data("SARS2003")
incid <- SARS2003$incidence
dt <- 7L
aggregate_inc(incid, dt)
}
