% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cdt_samples_convergence.R
\name{check_cdt_samples_convergence}
\alias{check_cdt_samples_convergence}
\title{Check MCMC chain convergence using the Gelman-Rubin algorithm}
\usage{
check_cdt_samples_convergence(cdt_samples)
}
\arguments{
\item{cdt_samples}{the \verb{@samples} slot of the output of
\code{coarseDataTools::dic.fit.mcmc()}}
}
\value{
TRUE if the Gelman Rubin test for convergence was successful, FALSE
otherwise
}
\description{
This function is deprecated as \code{\link[=estimate_R]{estimate_R()}} estimates the serial interval
by maximum likelihood. See \code{\link[=primary2estim]{primary2estim()}} for using primarycensored fits.
}
\details{
Splits an MCMC chain in two halves and uses the Gelman-Rubin algorithm to
assess convergence of the chain by comparing its two halves.
}
\seealso{
\code{\link[=primary2estim]{primary2estim()}}
}
\author{
Anne Cori
}
