% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cdt_samples_covergence.R
\name{check_cdtsamples_convergence}
\alias{check_cdtsamples_convergence}
\title{check_cdtsamples_convergence TITLE}
\usage{
check_cdtsamples_convergence(cdtsamples)
}
\arguments{
\item{cdtsamples}{XXXXXXX.}
}
\value{
TRUE if the Gelman Rubin test for convergence was successful, FALSE otherwise
}
\description{
\code{check_cdtsamples_convergence} DESCRIPTION TO COME.
}
\details{
{
XXXXXXX. 
}
}
\examples{
## XXXXXXX.
}
\references{
XXXXXXX.
}
\seealso{
XXXXXXX.
}
\author{
XXXXXXX.
}
