% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarse2estim.R
\name{coarse2estim}
\alias{coarse2estim}
\title{Link coarseDataTools and EpiEstim}
\usage{
coarse2estim(x = NULL, dist = x@dist, samples = x@samples, thin = 10)
}
\arguments{
\item{x}{An object generated by \code{coarseDataTools::dic.fit.mcmc()},
containing posterior estimates of the serial interval distribution.}

\item{dist}{The parametric distribution used when estimating the serial
interval. Should be one of "gamma", "weibull", "lognormal", "gamma_offset_1",
"weibull_offset_1", or "lognormal_offset_1". If not present, computed
automatically from \code{x}.}

\item{samples}{A dataframe containing the posterior samples of serial
interval parameters corresponding to the parametric choice specified in
\code{dist}. If not present, computed automatically from \code{x}.}

\item{thin}{A positive integer corresponding to thinning parameter; of the
posterior sample of serial interval distributions in x, only 1 in \code{thin}
will be kept, the rest will be discarded.}
}
\value{
A list with two elements:
\itemize{
\item \code{si_sample}: a matrix where each column gives one distribution of the
serial interval to be explored, obtained from x by thinning the MCMC chain.
\item \code{si_parametric_distr}: the parametric distribution used when estimating
the serial interval stored in x.
}
}
\description{
This function is deprecated. Use \code{\link[=primary2estim]{primary2estim()}}, which turns a serial
interval estimated with
\link[primarycensored:primarycensored-package]{primarycensored} into a sample
of serial interval distributions for \code{\link[=estimate_R]{estimate_R()}}.
}
\details{
Transforms the output of \code{coarseDataTools::dic.fit.mcmc()} to the right
format for input into \code{\link[=estimate_R]{estimate_R()}} with method "si_from_sample".
}
\seealso{
\code{\link[=primary2estim]{primary2estim()}}
}
\author{
The Hackout3 Parameter Estimation team.
}
