\encoding{UTF-8}

\name{coarse2estim}
\alias{coarse2estim}
\title{XXXX}
\description{
  \code{coarse2estim} XXXX. 
}
\usage{
coarse2estim(object, n_samples=1000)
}
\arguments{
  \item{object}{XXX}
  \item{n_samples}{XXX}
}
\value{
	XXX
}
\seealso{XXX}
\references{
  XXX.
}
\details{
  XXX
}
\author{XXX}
\examples{
# XXX
}
\keyword{XXX}
