% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarse2estim.R
\name{coarse2estim}
\alias{coarse2estim}
\title{coarse2estim TITLE}
\usage{
coarse2estim(object, n_samples = 1000)
}
\arguments{
\item{object}{XXXXXXX.}

\item{n_samples}{XXXXXXX..}
}
\value{
XXXXXXX.
}
\description{
\code{coarse2estim} DESCRIPTIOPN.
}
\details{
{
XXXXXXX. 
}
}
\examples{
## XXXXXXX.
}
\author{
XXXXXXX.
}
\references{
XXXXXXX.
}
\seealso{
XXXXXXX.
}

