% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{DiscrSI}
\alias{DiscrSI}
\title{Function to ensure compatibility with EpiEstim versions <2.0}
\usage{
DiscrSI(k, mu, sigma)
}
\arguments{
\item{k}{see \code{k} in \code{discr_si}}

\item{mu}{see \code{mu} in \code{discr_si}}

\item{sigma}{see \code{sigma} in \code{discr_si}}
}
\description{
Please only use for compatibility;
Prefer the new discr_si function instead
}
