% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_agg.R
\name{estimate_R_agg}
\alias{estimate_R_agg}
\title{Estimated Instantaneous Reproduction Number from coarsely aggregated data}
\usage{
estimate_R_agg(
  incid,
  dt = 7,
  dt_out = 7,
  iter = 10,
  config = make_config(),
  method = c("non_parametric_si", "parametric_si"),
  grid = list(precision = 0.001, min = -1, max = 1)
)
}
\arguments{
\item{incid}{aggregated incidence data, supplied as a vector}

\item{dt}{length of temporal aggregation of the data (numeric, 7 time units (typically days) by default)}

\item{dt_out}{length of the sliding windows for R estimates (numeric, 7 time units (typically days) by default)}

\item{iter}{number of iterations of the EM algorithm (numeric, 10 by default)}

\item{config}{An object of class \code{estimate_R_config}, as returned by 
function \code{make_config}.}

\item{method}{One of "non_parametric_si" or "parametric_si" (see details).}

\item{grid}{named list containing "precision", "min", and "max" which are used to
define a grid of growth rate parameters that are used inside the EM algorithm (see details)}
}
\value{
{
an object of class \code{estimate_R}, with components:
\itemize{

\item{R}{: a dataframe containing:
the times of start and end of each time window considered ;
the posterior mean, std, and 0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975
quantiles of the reproduction number for each time window.}

\item{method}{: the method used to estimate R, one of "non_parametric_si",
"parametric_si", "uncertain_si", "si_from_data" or "si_from_sample"}

\item{si_distr}{: a vector or dataframe (depending on the method) containing
 the discrete serial interval distribution(s) used for estimation}

\item{SI.Moments}{: a vector or dataframe (depending on the method)
containing the mean and std of the discrete serial interval distribution(s)
used for estimation}

\item{I}{: the time series of total incidence}

\item{I_local}{: the time series of incidence of local cases (so that
\code{I_local + I_imported = I})}

\item{I_imported}{: the time series of incidence of imported cases (so that
\code{I_local + I_imported = I})}

\item{dates}{: a vector of dates corresponding to the incidence time series}
}
}
}
\description{
Estimated Instantaneous Reproduction Number from coarsely aggregated data
}
\details{
TODO: add details
}
\examples{
## load data on SARS in 2003
data("SARS2003")

## aggregate the data to weekly counts of cases
incid <- SARS2003$incidence
dt <- 7L
weekly_inc <- aggregate_inc(incid, dt)
si_distr <- SARS2003$si_distr

## estimate Rt using the default parameters (method "non_parametric_si")
method <- "non_parametric_si"
config <- make_config(list(si_distr = si_distr))
res_weekly <- estimate_R_agg(incid = weekly_inc, 
                            dt = 7, # aggregation window of the data
                            dt_out = 7, # desired sliding window length
                            iter = 10,
                            config = config,
                            method = method,
                            grid = list(precision = 0.001, min = -1, max = 1))

## plot the result
plot(res_weekly)
}
