% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{EstimateR}
\alias{EstimateR}
\title{Title ### PLEASE USE THIS ONLY FOR COMPATIBILITY...}
\usage{
EstimateR(I, T.Start, T.End, method = c("NonParametricSI", "ParametricSI",
  "UncertainSI", "SIFromData", "SIFromSample"), n1 = NULL, n2 = NULL,
  Mean.SI = NULL, Std.SI = NULL, Std.Mean.SI = NULL, Min.Mean.SI = NULL,
  Max.Mean.SI = NULL, Std.Std.SI = NULL, Min.Std.SI = NULL,
  Max.Std.SI = NULL, SI.Distr = NULL, SI.Data = NULL,
  SI.parametricDistr = c("G", "W", "L", "off1G", "off1W", "off1L"),
  MCMC.control = list(init.pars = NULL, burnin = 3000, thin = 10, seed =
  as.integer(Sys.time())), SI.Sample = NULL, seed = NULL, Mean.Prior = 5,
  Std.Prior = 5, CV.Posterior = 0.3, plot = FALSE, legend = FALSE)
}
\arguments{
\item{I}{see I in \code{estimate_r}}

\item{T.Start}{see \code{config$t_start} in \code{estimate_r}}

\item{T.End}{.}

\item{method}{.}

\item{n1}{.}

\item{n2}{.}

\item{Mean.SI}{.}

\item{Std.SI}{.}

\item{Std.Mean.SI}{.}

\item{Min.Mean.SI}{.}

\item{Max.Mean.SI}{.}

\item{Std.Std.SI}{.}

\item{Min.Std.SI}{.}

\item{Max.Std.SI}{.}

\item{SI.Distr}{.}

\item{SI.Data}{.}

\item{SI.parametricDistr}{.}

\item{MCMC.control}{.}

\item{SI.Sample}{.}

\item{seed}{.}

\item{Mean.Prior}{.}

\item{Std.Prior}{.}

\item{CV.Posterior}{.}

\item{plot}{.}

\item{legend}{.}
}
\description{
Title ### PLEASE USE THIS ONLY FOR COMPATIBILITY...
}
