% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationR.R
\name{EstimateR}
\alias{EstimateR}
\title{Estimated Instantaneous Reproduction Number}
\usage{
EstimateR(I, T.Start, T.End, method = c("NonParametricSI", "ParametricSI",
  "UncertainSI", "SIFromData", "SIFromSample"), n1 = NULL, n2 = NULL,
  Mean.SI = NULL, Std.SI = NULL, Std.Mean.SI = NULL, Min.Mean.SI = NULL,
  Max.Mean.SI = NULL, Std.Std.SI = NULL, Min.Std.SI = NULL,
  Max.Std.SI = NULL, SI.Distr = NULL, SI.Data = NULL,
  SI.parametricDistr = c("G", "E", "off1G", "W", "L"),
  MCMC.control = list(init.pars = NULL, burnin = 3000), SI.Sample = NULL,
  Mean.Prior = 5, Std.Prior = 5, CV.Posterior = 0.3, plot = FALSE,
  leg.pos = "topright")
}
\arguments{
\item{I}{One of the following
\itemize{
\item{A vector of non-negative integers containing the incidence time series}
\item{A dataframe of non-negative integers with two columns, so that \code{I$local} contains the incidence of cases due to local transmission and \code{I$imported} contains the incidence of imported cases (with \code{I$local + I$imported} the total incidence).}
}}

\item{T.Start}{Vector of positive integers giving the starting times of each window over which the reproduction number will be estimated. These must be in ascending order, and so that for all \code{i}, \code{T.Start[i]<=T.End[i]}. T.Start[1] should be strictly after the first day with non null incidence.}

\item{T.End}{Vector of positive integers giving the ending times of each window over which the reproduction number will be estimated. These must be in ascending order, and so that for all \code{i}, \code{T.Start[i]<=T.End[i]}.}

\item{method}{Oone of "NonParametricSI", "ParametricSI", "UncertainSI", "SIFromData" or "SIFromSample" (see details).}

\item{n1}{For method "UncertainSI" and "SIFromData"; positive integer giving the size of the sample of SI distributions to be drawn (see details).}

\item{n2}{For methods "UncertainSI", "SIFromData" and "SIFromSample"; positive integer giving the size of the sample drawn from the posterior distribution of R for each serial interval distribution considered (see details).}

\item{Mean.SI}{For method "ParametricSI" and "UncertainSI" ; positive real giving the mean serial interval (method "ParametricSI") or the average mean serial interval (method "UncertainSI", see details).}

\item{Std.SI}{For method "ParametricSI" and "UncertainSI" ; non negative real giving the stadard deviation of the serial interval (method "ParametricSI") or the average standard deviation of the serial interval (method "UncertainSI", see details).}

\item{Std.Mean.SI}{For method "UncertainSI" ; standard deviation of the distribution from which mean serial intervals are drawn (see details).}

\item{Min.Mean.SI}{For method "UncertainSI" ; lower bound of the distribution from which mean serial intervals are drawn (see details).}

\item{Max.Mean.SI}{For method "UncertainSI" ; upper bound of the distribution from which mean serial intervals are drawn (see details).}

\item{Std.Std.SI}{For method "UncertainSI" ; standard deviation of the distribution from which standard deviations of the serial interval are drawn (see details).}

\item{Min.Std.SI}{For method "UncertainSI" ; lower bound of the distribution from which standard deviations of the serial interval are drawn (see details).}

\item{Max.Std.SI}{For method "UncertainSI" ; upper bound of the distribution from which standard deviations of the serial interval are drawn (see details).}

\item{SI.Distr}{For method "NonParametricSI" ; vector of probabilities giving the discrete distribution of the serial interval, starting with \code{SI.Distr[1]} (probability that the serial interval is zero), which should be zero.}

\item{SI.Data}{For method "SIFromData" ; the data on dates of symptoms of pairs of infector/infected individuals to be used to estimate the serial interval distribution (see details).}

\item{SI.parametricDistr}{For method "SIFromData" ; the parametric distribution to use when estimating the serial interval from data on dates of symptoms of pairs of infector/infected individuals (see details). 
Should be one of "G" (Gamma), "E" (Erlang), "off1G" (Gamma shifted by 1), "W" (Weibull), or "L" (Lognormal).}

\item{MCMC.control}{For method "SIFromData" ; a list containing the following (see details):
\describe{
  \item{init.pars}{vector of size 2 corresponding to the initial values of parameters to use for the SI distribution. This is the shape and scale for all but the lognormal distribution, for which it is the meanlog and sdlog. If not specified these are chosen automatically using function \code{\link{init_MCMC_params}}.}
  \item{burnin}{a positive integer giving the burnin used in the MCMC when estimating the serial interval distribution.}
}}

\item{SI.Sample}{For method "SIFromSample" ; a matrix where each column gives one distribution of the serial interval to be explored (see details).}

\item{Mean.Prior}{A positive number giving the mean of the common prior distribution for all reproduction numbers (see details).}

\item{Std.Prior}{A positive number giving the standard deviation of the common prior distribution for all reproduction numbers (see details).}

\item{CV.Posterior}{A positive number giving the aimed posterior coefficient of variation (see details).}

\item{plot}{Logical. If \code{TRUE} (default is \code{FALSE}), output is plotted (see value).}

\item{leg.pos}{One of "\code{bottomright}", "\code{bottom}", "\code{bottomleft}", "\code{left}", "\code{topleft}", "\code{top}", "\code{topright}", "\code{right}", "\code{center}" or \code{\link{xy.coords}(x, y)}, with \code{x} and \code{y} real numbers.}
}
\value{
{
a list with components: 
\itemize{
\item{R}{: a dataframe containing: 
the times of start and end of each time window considered ; 
the posterior mean, std, and 0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975 quantiles of the reproduction number for each time window.}
\item{SIDistr}{: a dataframe containing: 
for method "NonParametricSI", the mean and standard deviation of the discrete serial interval distribution;
for method "ParametricSI", the discrete serial interval distribution;
for method "UncertainSI", the means and standard deviations of the serial interval sampled to account for uncertainty on the serial interval distribution (see details);
for method "SIFromData", the means and standard deviations of the serial interval sampled from the posterior disrtribution obtained by Bayesian estimation from doubly censored data (see details).}
	}
	}
}
\description{
\code{EstimateR} estimates the reproduction number of an epidemic, given the incidence time series and the serial interval distribution.
}
\details{
{
Analytical estimates of the reproduction number for an epidemic over predefined time windows can be obtained within a Bayesian framework, 
for a given discrete distribution of the serial interval (see references). 

The more incident cases are observed over a time window, the smallest the posterior coefficient of variation (CV, ratio of standard deviation over mean) of the reproduction number. 
An aimed CV can be specified in the argument \code{CV.Posterior} (default is \code{0.3}), and a warning will be produced if the incidence within one of the time windows considered is too low to get this CV. 

The methods vary in the way the serial interval distribution is specified. The plots are also different according to the method used.

----------------------- \code{method "NonParametricSI"} -----------------------
  
The discrete distribution of the serial interval is directly specified in the argument \code{SI.Distr}.

If \code{plot} is \code{TRUE}, 3 plots are produced. 
The first one shows the epidemic curve. 
The second one shows the posterior median and 95\% credible interval of the reproduction number. The estimate for a time window is plotted at the end of the time window. 
The position of the legend on that graph can be monitored by the argument \code{leg.pos} (default is "\code{topright}").
The third plot shows the discrete distribution of the serial interval. 

----------------------- \code{method "ParametricSI"} -----------------------
  
The mean and standard deviation of the continuous distribution of the serial interval are given in the arguments \code{Mean.SI} and \code{Std.SI}.
The discrete distribution of the serial interval is derived automatically using \code{\link{DiscrSI}}.

If \code{plot} is \code{TRUE}, 3 plots are produced, which are identical to the ones for \code{method "NonParametricSI"} .

----------------------- \code{method "UncertainSI"} -----------------------
   
\code{Method "UncertainSI"} allows accounting for uncertainty on the serial interval distribution (see Cori et al. AJE 2013). 
We allow the mean \eqn{\mu} and standard deviation \eqn{\sigma} of the serial interval to vary according to truncated normal distributions. 
We sample \code{n1} pairs of mean and standard deviations, \eqn{(\mu^{(1)},\sigma^{(1)}),...,(\mu^{(n_2)},\sigma^{(n_2)})}, by first sampling the mean \eqn{\mu^{(k)}} 
from its truncated normal distribution (with mean \code{Mean.SI}, standard deviation \code{Std.Mean.SI}, minimum \code{Min.Mean.SI} and maximum \code{Max.Mean.SI}), 
and then sampling the standard deviation \eqn{\sigma^{(k)}} from its truncated normal distribution 
(with mean \code{Std.SI}, standard deviation \code{Std.Std.SI}, minimum \code{Min.Std.SI} and maximum \code{Max.Std.SI}), but imposing that \eqn{\sigma^{(k)}<\mu^{(k)}}. 
This constraint ensures that the Gamma probability density function of the serial interval is null at \eqn{t=0}. 
Warnings are produced when the truncated normal distributions are not symmetric around the mean. 
For each pair \eqn{(\mu^{(k)},\sigma^{(k)})}, we then draw a sample of size \code{n2} in the posterior distribution of the reproduction number over each time window, conditionnally on this serial interval distribution. 
After pooling, a sample of size \eqn{\code{n1}\times\code{n2}} of the joint posterior distribution of the reproduction number over each time window is obtained.
The posterior mean, standard deviation, and 0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975 quantiles of the reproduction number for each time window are obtained from this sample.

If \code{plot} is \code{TRUE}, 4 plots are produced.
The first one shows the epidemic curve. 
The second one shows the posterior median and 95\% credible interval of the reproduction number. The estimate for a time window is plotted at the end of the time window. 
The position of the legend on that graph can be monitored by the argument \code{leg.pos} (default is "\code{topright}").
The third and fourth plots show histograms of the sampled means and standard deviations of the serial interval. 

----------------------- \code{method "SIFromData"} -----------------------
  
\code{Method "SIFromData"} allows accounting for uncertainty on the serial interval distribution. 
Unlike method "UncertainSI", where we arbitrarily vary the mean and std of the SI in truncated normal distributions, 
here, the scope of serial interval distributions considered is directly informed by data
on the (potentially censored) dates of symptoms of pairs of infector/infected individuals (specified in XXX TO BE COMPLETED XXXX). 
Assuming a given parametric distribution for the serial interval distribution (specified in XXX TO BE COMPLETED XXXX), 
the posterior distribution of the serial interval is estimated directly fom these data using MCMC methods implemented in the package \code{coarsedatatools}. 
XXX NEED TO ADD SOME OPTIONS FOR THE MCMC SUCH AS BURNIN, NUMBER OF ITERATIONS AND HENCE POSTERIOR SAMPLE SIZE ETC. XXX
For each element in the posterior sample of serial interval distribution, we draw a sample of size \code{n2} in the posterior distribution of the reproduction number over each time window, conditionnally on this serial interval distribution. 
After pooling, a sample of size \eqn{\code{XXX TO BE COMPLETED XXX}\times\code{n2}} of the joint posterior distribution of the reproduction number over each time window is obtained.
The posterior mean, standard deviation, and 0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975 quantiles of the reproduction number for each time window are obtained from this sample.

If \code{plot} is \code{TRUE}, 4 plots are produced.
The first one shows the epidemic curve. 
The second one shows the posterior median and 95\% credible interval of the reproduction number. The estimate for a time window is plotted at the end of the time window. 
The position of the legend on that graph can be monitored by the argument \code{leg.pos} (default is "\code{topright}").
The third and fourth plots show histograms of the sampled means and standard deviations of the serial interval. 

#' ----------------------- \code{method "SIFromSample"} -----------------------
XXX TO BE COMPLETED XXX
}
}
\examples{
## load data on pandemic flu in a school in 2009
data("Flu2009")

## estimate the reproduction number (method "NonParametricSI")
EstimateR(Flu2009$Incidence, T.Start=2:26, T.End=8:32, method="NonParametricSI", 
          SI.Distr=Flu2009$SI.Distr, plot=TRUE, leg.pos=xy.coords(1,3))
# the second plot produced shows, at each each day, 
# the estimate of the reproduction number over the 7-day window finishing on that day.

## estimate the reproduction number (method "ParametricSI")
EstimateR(Flu2009$Incidence, T.Start=2:26, T.End=8:32, method="ParametricSI", 
          Mean.SI=2.6, Std.SI=1.5, plot=TRUE)
# the second plot produced shows, at each each day, 
# the estimate of the reproduction number over the 7-day window finishing on that day.

## estimate the reproduction number (method "UncertainSI")
EstimateR(Flu2009$Incidence, T.Start=2:26, T.End=8:32, method="UncertainSI", 
          Mean.SI=2.6, Std.Mean.SI=1, Min.Mean.SI=1, Max.Mean.SI=4.2, 
          Std.SI=1.5, Std.Std.SI=0.5, Min.Std.SI=0.5, Max.Std.SI=2.5, 
          n1=100, n2=100, plot=TRUE)
# the bottom left plot produced shows, at each each day, 
# the estimate of the reproduction number over the 7-day window finishing on that day.

\dontrun{
## Note the following examples use an MCMC routine 
## to estimate the serial interval distribution from data, 
## so they may take a few minutes to run

## load data on rotavirus
data("MockRotavirus")

## estimate the reproduction number (method "SIFromData")
set.seed(1)
R_SIFromData <- EstimateR(MockRotavirus$Incidence, 
                                        T.Start=2:47, T.End=8:53, 
                                        method="SIFromData", 
                                        SI.Data=MockRotavirus$SI.Data, 
                                        SI.parametricDistr = "G", 
                                        MCMC.control = list(burnin = 1000), 
                                        n1 = 1000, n2 = 50,
                                        plot=TRUE, leg.pos=xy.coords(1,3))
## compare with version with no uncertainty
R_Parametric <- EstimateR(MockRotavirus$Incidence, 
                          T.Start=2:47, T.End=8:53, 
                          method="ParametricSI", 
                          Mean.SI = mean(R_SIFromData$SIDistr$Mean.SI.sample), 
                          Std.SI = mean(R_SIFromData$SIDistr$Std.SI.sample), 
                          plot=TRUE)
## generate plots
p_uncertainty <- plots(R_SIFromData, "R")
p_no_uncertainty <- plots(R_Parametric, "R")
gridExtra::grid.arrange(p_uncertainty, p_no_uncertainty,ncol=2)
# the left hand side graph is with uncertainty in the SI distribution, the right hand side without. 
# The credible intervals are wider when accounting for uncertainty in the SI distribution. 

#' ## estimate the reproduction number (method "SIFromSample")
set.seed(1)
SI.fit <- coarseDataTools::dic.fit.mcmc(dat = MockRotavirus$SI.Data, 
                             dist="G", 
                             init.pars=init_MCMC_params(MockRotavirus$SI.Data, "G")
                             burnin = 1000, 
                             n.samples = 1000)
SI.Sample <- coarse2estim(SI.fit, nrow(SI.fit@samples))$prob_matrix
R_SIFromSample <- EstimateR(MockRotavirus$Incidence, 
                            T.Start=2:47, T.End=8:53, 
                            method="SIFromSample", SI.Sample=SI.Sample,
                            n2 = 50,
                            plot=TRUE, leg.pos=xy.coords(1,3))

# check that R_SIFromSample is the same as R_SIFromData 
# since they were generated using the same MCMC algorithm to generate the SI sample
# (either internally to EpiEstim or externally)
all(R_SIFromSample$R$`Mean(R)` == R_SIFromData$R$`Mean(R)`) 
}

}
\author{
Anne Cori \email{a.cori@imperial.ac.uk}
}
\references{
{
Cori, A. et al. A new framework and software to estimate time-varying reproduction numbers during epidemics (AJE 2013).
Wallinga, J. and P. Teunis. Different epidemic curves for severe acute respiratory syndrome reveal similar impacts of control measures (AJE 2004).
}
}
\seealso{
\code{\link{DiscrSI}}
}

