% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_mcmc_params.R
\name{init_mcmc_params}
\alias{init_mcmc_params}
\title{Find starting points for the estimation of the serial interval}
\usage{
init_mcmc_params(si_data, dist)
}
\arguments{
\item{si_data}{The data on dates of symptoms of pairs of infector/infected
individuals, as described in \code{\link[=estimate_R]{estimate_R()}}.}

\item{dist}{The parametric distribution of the serial interval, as in
\code{si_parametric_distr} in \code{\link[=make_config]{make_config()}}, e.g. "gamma" or
"gamma_offset_1".}
}
\value{
A vector containing the starting values for the parameters of the
distribution of the serial interval, as given by \code{\link[=si_start_values]{si_start_values()}}.
}
\description{
This function is deprecated. Use \code{\link[=si_start_values]{si_start_values()}} instead.
}
\seealso{
\code{\link[=si_start_values]{si_start_values()}}
}
\author{
Anne Cori
}
