% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_MCMC_params.R
\name{init_MCMC_params}
\alias{init_MCMC_params}
\title{init_MCMC_params TITLE}
\usage{
init_MCMC_params(si_data, dist = c("G", "W", "L", "off1G", "off1W", "off1L"))
}
\arguments{
\item{si_data}{XXXXXXX.}

\item{dist}{XXXXXXX.}
}
\value{
XXXXXXX.
}
\description{
\code{init_MCMC_params} DESCRIPTION TO COME.
}
\details{
{
XXXXXXX. 
}
}
\examples{
## XXXXXXX.
}
\references{
XXXXXXX.
}
\seealso{
XXXXXXX.
}
\author{
XXXXXXX.
}
