% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mcmc_control.R
\name{make_mcmc_control}
\alias{make_mcmc_control}
\title{Create list of MCMC control parameters}
\usage{
make_mcmc_control(
  burnin = 3000,
  thin = 10,
  seed = as.integer(Sys.time()),
  init_pars = NULL
)
}
\arguments{
\item{burnin}{Not used.}

\item{thin}{Not used.}

\item{seed}{An integer used as the seed for the random number generator
when drawing the sample of serial interval distributions.}

\item{init_pars}{Starting values for the parameters of the serial interval
distribution, as given by \code{\link[=si_start_values]{si_start_values()}}.}
}
\value{
An object of class \code{estimate_R_mcmc_control} with components
\code{burnin}, \code{thin}, \code{seed}, \code{init_pars}. This can be
used as an argument of function \code{\link[=make_config]{make_config()}}.
}
\description{
This function is deprecated. The serial interval is estimated by maximum
likelihood in \code{\link[=estimate_R]{estimate_R()}} with method "si_from_data", so \code{burnin} and
\code{thin} are not used, and a warning is given if they differ from their
defaults. Use \code{seed} in \code{\link[=make_config]{make_config()}} to make the sample of serial
interval distributions reproducible.
}
\seealso{
\code{\link[=si_start_values]{si_start_values()}}
}
