% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverallInfectivity.R
\name{OverallInfectivity}
\alias{OverallInfectivity}
\title{Overall Infectivity Due To Previously Infected Individuals}
\usage{
OverallInfectivity(I, SI.Distr)
}
\arguments{
\item{I}{One of the following
\itemize{
\item{A vector of non-negative integers containing an incidence time series}
\item{A dataframe of non-negative integers with two columns, so that \code{I$local} contains the incidence of cases due to local transmission and \code{I$imported} contains the incidence of imported cases (with \code{I$local + I$imported} the total incidence).}
} 
Note that the cases from the first time step are always all assumed to be imported cases.}

\item{SI.Distr}{Vector of probabilities giving the discrete distribution of the serial interval.}
}
\value{
A vector which contains the overall infectivity \eqn{\lambda_t} at each time step
}
\description{
\code{OverallInfectivity} computes the overall infectivity due to previously infected individuals.
}
\details{
{
The overall infectivity \eqn{\lambda_t} at time step \eqn{t} is equal to the sum of the previously infected individuals 
(given by the incidence vector \eqn{I}, with \code{I=I$local + I$imported} if \eqn{I} is a matrix), 
weigthed by their infectivity at time \eqn{t} (given by the discrete serial interval distribution \eqn{w_k}). 
In mathematical terms:   
\cr
\eqn{\lambda_t = \sum_{k=1}^{t-1}I_{t-k}w_k}
\cr
}
}
\examples{
## load data on pandemic flu in a school in 2009
data("Flu2009")

## compute overall infectivity
lambda <- OverallInfectivity(Flu2009$Incidence, Flu2009$SI.Distr)
par(mfrow=c(2,1))
plot(Flu2009$Incidence, type="s", xlab="time (days)", ylab="Incidence")
title(main="Epidemic curve")
plot(lambda, type="s", xlab="time (days)", ylab="Infectivity")
title(main="Overall infectivity")
}
\author{
Anne Cori \email{a.cori@imperial.ac.uk}
}
\references{
Cori, A. et al. A new framework and software to estimate time-varying reproduction numbers during epidemics (AJE 2013).
}
\seealso{
\code{\link{DiscrSI}}, \code{\link{EstimateR}}
}

