% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{OverallInfectivity}
\alias{OverallInfectivity}
\title{Function to ensure compatibility with EpiEstim versions <2.0}
\usage{
OverallInfectivity(I, SI.Distr)
}
\arguments{
\item{I}{see \code{incid} in \code{overall_infectivity}}

\item{SI.Distr}{see \code{si_distr} in \code{overall_infectivity}}
}
\description{
Please only use for compatibility;
Prefer the new overall_infectivity function instead
}
