% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\title{Plotting the outputs of functions estimating the reproduction number from incidence time series and assumptions regarding the serial interval distribution}
\usage{
plots(x = NULL, what = c("all", "I", "R", "SI"), x2 = NULL,
  add_imported_cases = FALSE, ylim = NULL, options_SI = list(prob_min =
  0.001, transp = 0.25))
}
\arguments{
\item{x}{The output of function \code{\link{EstimateR}} or function \code{\link{WT}}}

\item{what}{A string specifying what to plot, namely the incidence time series (\code{what='I'}), the estimated reproduction number (\code{what='R'}), the serial interval distribution (\code{what='SI'}, or all three (\code{what='all'})).}

\item{x2}{Optional, and only used if \code{what='R'} or \code{what='all'}. A second output of function \code{\link{EstimateR}} or function \code{\link{WT}}, 
typically based on the same incidence, for which the Reproduction number plot will be overlayed on the other one}

\item{add_imported_cases}{A boolean to specify whether, on the incidence time series plot, to add the incidence of imported cases.}

\item{ylim}{For what = "I" or "R"; a parameter similar to that in \code{par}, to monitor the limits of the vertical axis}

\item{options_SI}{For what = "SI". A list of graphical options: 
 \describe{
\item{prob_min}{A numeric value between 0 and 1. The SI distributions explored are only shown from time 0 up to the time t so that each distribution explored has probability < \code{prob_min} to be on any time step after t. Defaults to 0.001.}
\item{transp}{A numeric value between 0 and 1 used to monitor transparency of the lines. Defaults to 0.25}
}}
}
\value{
a plot (if \code{what = "I"}, \code{"R"}, or \code{"SI"}) or a \code{\link{grob}} object (if \code{what = "all"}).
}
\description{
\code{plots} allows plotting the outputs of functions \code{\link{EstimateR}} and \code{\link{WT}}
}
\examples{
## load data on pandemic flu in a school in 2009
data("Flu2009")

## estimate the instantaneous reproduction number (method "NonParametricSI")
R_i <- EstimateR(Flu2009$Incidence, T.Start=2:26, T.End=8:32, method="NonParametricSI", 
          SI.Distr=Flu2009$SI.Distr, plot=FALSE)

## visualise results
plots(R_i)

## estimate the instantaneous reproduction number (method "NonParametricSI")
R_c <- WT(Flu2009$Incidence, T.Start=2:26, T.End=8:32, method="NonParametricSI", 
          SI.Distr=Flu2009$SI.Distr, plot=FALSE)

## produce plot of the incidence 
       ## (with, on top of total incidence, the incidence of imported cases), 
       ## estimated instantaneous and case reproduction numbers 
       ## and serial interval distribution used
p_I <- plots(R_i, "I", add_imported_cases=TRUE) # plots the incidence 
p_SI <- plots(R_i, "SI") # plots the serial interval distribution
p_Ri <- plots(R_i, "R", ylim=c(0,4)) # plots the estimated instantaneous reproduction number
p_Rc <- plots(R_c, "R", ylim=c(0,4)) # plots the estimated case reproduction number
gridExtra::grid.arrange(p_I,p_SI,p_Ri,p_Rc,ncol=2)

}
\seealso{
\code{\link{EstimateR}} and \code{\link{WT}}
}
\author{
Rolina van Gaalen \email{rolina.van.gaalen@rivm.nl} and Anne Cori \email{a.cori@imperial.ac.uk}
}
