% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{si_from_data_valid_distrs}
\alias{si_from_data_valid_distrs}
\title{Distribution names valid when using MCMC to estimate SI from data}
\usage{
si_from_data_valid_distrs(dist)
}
\arguments{
\item{dist}{The parametric distribution used when estimating the serial
interval. Should be one of "gamma", "weibull", "lognormal", "gamma_offset_1",
 "weibull_offset_1", or "lognormal_offset_1". Note the different naming 
convention compared to \code{\link[coarseDataTools]{dic.fit.mcmc}}. The
distribution may also be specified using the abbreviated forms "G", "W", "L",
"off1G", "off1W", and "off1L" as used in \code{\link[coarseDataTools]{dic.fit.mcmc}}.
However, we recommend using the full names to avoid confusion, and a warning
will be issued if the abbreviated forms are used.
If not present, computed automatically from \code{x}.}
}
\value{
A two element list - the first element is a flag `is_dist_valid`
indicating whether the passed distribution is valid. Te second element
`all_valid_distrs` is a character vector with the valid distribution names.
}
\description{
When using si_from_data method, the package will use
\code{\link[coarseDataTools]{dic.fit.mcmc}} to fit the serial interval
distribution. This method supports only a limited set of distributions.
This function returns the valid distribution names as used by EpiEstim. The
names are internally converted to those used by coarsedatatools by
\code{\link{convert_distr_name_for_mcmc}} function.
}
\author{
Sangeeta Bhatia
}
