% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{si_from_data_valid_distrs}
\alias{si_from_data_valid_distrs}
\title{Distribution names valid when using MCMC to estimate SI from data}
\usage{
si_from_data_valid_distrs()
}
\value{
A character vector with the valid distribution names.
}
\description{
When using si_from_data method, the package will use
\code{\link[coarseDataTools]{dic.fit.mcmc}} to fit the serial interval
distribution. This method supports only a limited set of distributions.
This function returns the valid distribution names as used by EpiEstim. The
names are internally converted to those used by coarsedatatools by
\code{\link{convert_distr_name_for_mcmc}} function.
}
\author{
Sangeeta Bhatia
}
