% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{WT}
\alias{WT}
\title{Function to ensure compatibility with EpiEstim versions <2.0}
\usage{
WT(
  I,
  T.Start,
  T.End,
  method = c("NonParametricSI", "ParametricSI"),
  Mean.SI = NULL,
  Std.SI = NULL,
  SI.Distr = NULL,
  nSim = 10,
  plot = FALSE,
  leg.pos = "topright"
)
}
\arguments{
\item{I}{see \code{incid} in \code{wallinga_teunis}}

\item{T.Start}{see \code{config$t_start} in \code{wallinga_teunis}}

\item{T.End}{see \code{config$t_end} in \code{wallinga_teunis}}

\item{method}{see method in \code{wallinga_teunis} (but WT uses CamelCase 
where wallinga_teunis uses snake_case for the method names)}

\item{Mean.SI}{see \code{config$mean_si} in \code{wallinga_teunis}}

\item{Std.SI}{see \code{config$std_si} in \code{wallinga_teunis}}

\item{SI.Distr}{see \code{config$si_distr} in \code{wallinga_teunis}}

\item{nSim}{see \code{config$n_sim} in \code{wallinga_teunis}}

\item{plot}{Not used anymore, only there for compatibility}

\item{leg.pos}{Not used anymore, only there for compatibility}
}
\description{
Please only use for compatibility;
Prefer the new wallinga_teunis function instead
}
